Move client_builder to k8s.io/controller-manager
