+++ cluster/gce/gci/audit_policy_test.go
@@ -26,9 +26,9 @@ import (
 	auditinstall "k8s.io/apiserver/pkg/apis/audit/install"
 	auditpkg "k8s.io/apiserver/pkg/audit"
 	auditpolicy "k8s.io/apiserver/pkg/audit/policy"
+	"k8s.io/apiserver/pkg/authentication/serviceaccount"
 	"k8s.io/apiserver/pkg/authentication/user"
 	"k8s.io/apiserver/pkg/authorization/authorizer"
-	"k8s.io/kubernetes/pkg/serviceaccount"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
