+++ staging/src/k8s.io/controller-manager/BUILD
@@ -26,6 +26,7 @@ filegroup(
         "//staging/src/k8s.io/controller-manager/app:all-srcs",
         "//staging/src/k8s.io/controller-manager/config:all-srcs",
         "//staging/src/k8s.io/controller-manager/options:all-srcs",
+        "//staging/src/k8s.io/controller-manager/pkg/clientbuilder:all-srcs",
         "//staging/src/k8s.io/controller-manager/pkg/features:all-srcs",
     ],
     tags = ["automanaged"],
