+++ staging/src/k8s.io/controller-manager/go.mod
@@ -7,6 +7,7 @@ go 1.15
 require (
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.4.0
+	k8s.io/api v0.0.0
 	k8s.io/apimachinery v0.0.0
 	k8s.io/apiserver v0.0.0
 	k8s.io/client-go v0.0.0
